Many DB Changes

+++ DLL/Item.cs
@@ -5,7 +5,7 @@ namespace DLL
     public abstract class Item
     {
         public int Id { get; set; }
-        public int ItemCode { get; set; }
+        public long ItemCode { get; set; }
         public string Name { get; set; }
         public int Stock { get; set; }
         public double Price { get; set; }
