+++ DLL/Person.cs
@@ -7,7 +7,7 @@ namespace DLL
     public abstract class Person
     {
         public int Id { get; set; }
-        public int PersonStoreID { get; set; }
+        public long PersonStoreID { get; set; }
         public string FirstName { get; set; }
         public string LastName { get; set; }
         public DateTime? Birthdate { get; set; }
